apps/taxi/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
PyJWT==2.9.0
python-dotenv==1.0.1
prometheus-client==0.20.0
pytest==8.3.3
httpx==0.27.2
redis==5.0.8
alembic==1.13.2
sentry-sdk==2.14.0
paho-mqtt==2.1.0
